feat: Add storage-type (#47)
Sometime there is a need to specify storage (localStorage or Cookies), like in situations, when you use Nuxt.js and there is a need to read 'cookies:accept' in `nuxtServerInit` hook. So, you can't work with **localStorage** and with **SSR**. Thats why it will be great to choose where to store.

I've added prop called `storageType`. It can be 'localStorage' (default) or 'cookies'.